Добавлены правки по ревью
