+++ ImageFeed/Network/URLSession+data.swift
@@ -6,28 +6,28 @@ extension URLSession {
         completion: @escaping (Result<T, Error>) -> Void
     ) -> URLSessionTask {
         let decoder = JSONDecoder()
-//        decoder.keyDecodingStrategy = .convertFromSnakeCase
+        //        decoder.keyDecodingStrategy = .convertFromSnakeCase
         let task = data(for: request) { (result: Result<Data, Error>) in
-                switch result {
-                case .success(let data):
-                    do {
-                        let decodedObject = try decoder.decode(T.self, from: data)
-                        print("Decoding success")
-                        completion(.success(decodedObject))
-                    } catch {
-                        if let decodingError = error as? DecodingError {
-                            let authError = AuthServiceError.decodingError(decodingError)
-                            print("Ошибка декодирования: \(authError), Данные: \(String(data: data, encoding: .utf8) ?? "")")
-                        } else {
-                            print("Ошибка декодирования: \(error.localizedDescription), Данные: \(String(data: data, encoding: .utf8) ?? "")")
-                        }
-                        completion(.failure(error))
-                        
+            switch result {
+            case .success(let data):
+                do {
+                    let decodedObject = try decoder.decode(T.self, from: data)
+                    print("Decoding success")
+                    completion(.success(decodedObject))
+                } catch {
+                    if let decodingError = error as? DecodingError {
+                        let authError = AuthServiceError.decodingError(decodingError)
+                        print("Ошибка декодирования: \(authError), Данные: \(String(data: data, encoding: .utf8) ?? "")")
+                    } else {
+                        print("Ошибка декодирования: \(error.localizedDescription), Данные: \(String(data: data, encoding: .utf8) ?? "")")
                     }
-                case .failure(let error):
-                    print("Ошибка запроса: \(error.localizedDescription)")
                     completion(.failure(error))
+                    
                 }
+            case .failure(let error):
+                print("Ошибка запроса: \(error.localizedDescription)")
+                completion(.failure(error))
+            }
             
         }
         return task
