+++ ImageFeed/OAuth2/DecodindService.swift
@@ -1,0 +1,16 @@
+import UIKit
+
+struct OAuthTokenResponseBody: Decodable {
+    
+    let accessToken: String
+    let tokenType: String
+    let scope: String
+    let createdAt: Int
+    
+    private enum CodingKeys: String, CodingKey {
+        case accessToken = "access_token"
+        case tokenType = "token_type"
+        case scope
+        case createdAt = "created_at"
+    }
+}
