+++ ImageFeed/OAuth2/ErrorService.swift
@@ -1,0 +1,11 @@
+import UIKit
+
+enum AuthServiceError: Error {
+    case httpStatusCode(Int)
+    case urlRequestError(Error)
+    case urlSessionError
+    case invalidRequest
+    case decodingError(DecodingError)
+}
+
+
