+++ ImageFeed/OAuth2/OAuth2Service.swift
@@ -1,28 +1,5 @@
 import UIKit
 
-enum AuthServiceError: Error {
-    case httpStatusCode(Int)
-    case urlRequestError(Error)
-    case urlSessionError
-    case invalidRequest
-    case decodingError(DecodingError)
-}
-
-struct OAuthTokenResponseBody: Decodable {
-    
-    let accessToken: String
-    let tokenType: String
-    let scope: String
-    let createdAt: Int
-    
-    private enum CodingKeys: String, CodingKey {
-        case accessToken = "access_token"
-        case tokenType = "token_type"
-        case scope
-        case createdAt = "created_at"
-    }
-}
-
 final class OAuth2Service {
     private let urlSession = URLSession.shared
     private var task: URLSessionTask?
@@ -81,35 +58,31 @@ final class OAuth2Service {
         }
         self.task = task
         task.resume()
-        }
-            }
-        
-    
-    private func makeOAuthTokenRequest(code: String) -> URLRequest? {
-        guard var urlComponents = URLComponents(string: "https://unsplash.com/oauth/token") else {
-            print("Ошибка в URL Components")
-            return nil
-        }
-        urlComponents.queryItems = [
-            URLQueryItem(name: "client_id", value: Constants.accessKey),
-            URLQueryItem(name: "client_secret", value: Constants.secretKey),
-            URLQueryItem(name: "redirect_uri", value: Constants.redirectURI),
-            URLQueryItem(name: "code", value: code),
-            URLQueryItem(name: "grant_type", value: "authorization_code")
-        ]
-        
-        guard let authTokenUrl = urlComponents.url else {
-            print("Ошибка в создании URL с queryItems")
-            return nil
-        }
-        var request = URLRequest(url: authTokenUrl)
-        request.httpMethod = "POST"
-        return request
     }
+}
+
+
+private func makeOAuthTokenRequest(code: String) -> URLRequest? {
+    guard var urlComponents = URLComponents(string: AuthConstants.authForFetchTokenURLString) else {
+        print("Ошибка в URL Components")
+        return nil
+    }
+    urlComponents.queryItems = [
+        URLQueryItem(name: "client_id", value: Constants.accessKey),
+        URLQueryItem(name: "client_secret", value: Constants.secretKey),
+        URLQueryItem(name: "redirect_uri", value: Constants.redirectURI),
+        URLQueryItem(name: "code", value: code),
+        URLQueryItem(name: "grant_type", value: "authorization_code")
+    ]
     
-
-//    }
-
+    guard let authTokenUrl = urlComponents.url else {
+        print("Ошибка в создании URL с queryItems")
+        return nil
+    }
+    var request = URLRequest(url: authTokenUrl)
+    request.httpMethod = "POST"
+    return request
+}
