+++ ImageFeed/Views/UIBlockingProgressHUD.swift
@@ -15,5 +15,5 @@ final class UIBlockingProgressHUD {
         window?.isUserInteractionEnabled = true
         ProgressHUD.dismiss()
     }
-
+    
 }
